chore(compiler): Removing weird path-rewrite, defaulting to mjs for all entrypoint script
