+++ module/cli/bin/trv.js
@@ -1,2 +1,6 @@
 #!/usr/bin/env node
-require('./trv.mjs');
+
+// @ts-check
+const { getEntry } = require('@travetto/compiler/bin/common.js');
+
+getEntry().then(ops => ops.getLoader()).then(load => load('@travetto/cli/support/entry.trv.js'));
